Disgusting test code for multiple outline colors
I promise I'm cleaning this up tomorrow

With some more temporary code, successfully got multiple colors of outlines being drawn, and the manager deleting only certain kinds.

Immediate next task is to come up with better organization for the types of outlines, and rewriting the very placeholdery validate-and-do-action method in the placeholder.