Pop-up on meeting another player
Added a new delegate to the HexClickHandler and renamed it to simply HexDelegates, representing its generalized function of letting hexes call specific methods at the gamemaster level.

A panel has been added to the UI, and it and the UI have scripts that allow them to be accessed.

Players now track whether they have met any other given player. This is checked when determining whether or not to show this panel when a player is "sighted."

Tokens now know their owner.