Greeting pop-up
When first sighting the unit of another player, a greeting pop up is shown.

The pop-up panel now has a title and body text field that it can modify.

A language class was created. Will be fleshed out later, but for now just creates passable names for the players. This is shown in the greeting pop up.