Outline and OutlineManager overhaul
Drastically improved the implementation of Outlines and Outline managers

The most important change is that the manager does not create and destroy new outlines every single turn. Each hex now has two outlines, starting as null, that they keep. They are stored there. The outline manager still creates the outlines, but it now stores them on the hex and simply hides them when they are no longer used.

This also prevents a single hex from having multiple overlapping outlines of the same type.

Hexes that have other player's units in them are now shown with a red outline.